CMake Overhaul

+++ C/2022/day05.c
@@ -1,27 +1,31 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-char* one(FILE* fp)
+char *one(FILE *fp)
 {
     // prep the input
-    char* line = NULL;
+    char *line = NULL;
     ssize_t read;
     size_t len = 0;
 
     // read until the newline
     size_t des_line = 0;
     int num_stacks = 0;
-    while ((read = getline(&line, &len, fp)) != 1) {
+    while ((read = getline(&line, &len, fp)) != 1)
+    {
         // our desired line guaranteed starts with a space
-        if (line[0] == ' ') {
+        if (line[0] == ' ')
+        {
             // get the number of stacks
             size_t num_idx = 0;
-            char* num = malloc(3 * sizeof(char));
-            for (size_t idx = read - 2; line[idx] != ' '; --idx)
+            char *num = malloc(3 * sizeof(char));
+            for (int idx = read - 2; line[idx] != ' '; --idx)
                 num[num_idx++] = line[idx];
             num_stacks = atoi(num);
             break;
-        } else {
+        }
+        else
+        {
             ++des_line;
         }
     }
@@ -30,20 +34,23 @@ char* one(FILE* fp)
     fseek(fp, 0, 0);
 
     // allocate the storage
-    char** stacks = malloc(num_stacks * sizeof(char*));
-    for (size_t i = 0; i < num_stacks; ++i) {
+    char **stacks = malloc(num_stacks * sizeof(char *));
+    for (int i = 0; i < num_stacks; ++i)
+    {
         stacks[i] = malloc(100 * sizeof(char));
     }
-    size_t* size = malloc(9 * sizeof(size_t));
+    size_t *size = malloc(9 * sizeof(size_t));
 
     // now read the first 8 lines and get the data
     size_t line_num = 0;
-    while (line_num != des_line) {
+    while (line_num != des_line)
+    {
         // get the line
         read = getline(&line, &len, fp);
 
         // read across the line looking for chars
-        for (size_t i = 0; i < read; ++i) {
+        for (int i = 0; i < read; ++i)
+        {
             if (line[i] >= 'A' && line[i] <= 'Z')
                 stacks[i / 4][size[i / 4]++] = line[i];
         }
@@ -52,8 +59,9 @@ char* one(FILE* fp)
     }
 
     // we put them in backwards, reverse the lists
-    for (size_t i = 0; i < num_stacks; ++i) {
-        char* rev_stack = malloc(100 * sizeof(char));
+    for (int i = 0; i < num_stacks; ++i)
+    {
+        char *rev_stack = malloc(100 * sizeof(char));
         for (size_t j = 0; j <= size[i]; ++j)
             rev_stack[j] = stacks[i][size[i] - j - 1];
         free(stacks[i]);
@@ -66,7 +74,8 @@ char* one(FILE* fp)
     getline(&line, &len, fp);
 
     // now we need to parse the rest of the lines
-    while (getline(&line, &len, fp) != -1) {
+    while (getline(&line, &len, fp) != -1)
+    {
         // move to first number
         size_t idx;
         for (idx = 1; line[idx - 1] != ' '; ++idx)
@@ -74,7 +83,7 @@ char* one(FILE* fp)
 
         // get first number
         size_t num_idx = 0;
-        char* num = malloc(3 * sizeof(char));
+        char *num = malloc(3 * sizeof(char));
         for (; line[idx] != ' '; ++idx)
             num[num_idx++] = line[idx];
         const size_t moving = atoi(num);
@@ -104,18 +113,19 @@ char* one(FILE* fp)
         const size_t destination = atoi(num) - 1;
 
         // perform the action
-        for (size_t action = 0; action < moving; ++action) {
+        for (size_t action = 0; action < moving; ++action)
+        {
             stacks[destination][size[destination]++] = stacks[origin][(size[origin]--) - 1];
         }
     }
 
     // store top stacks in return
-    char* ret = malloc(num_stacks * sizeof(char));
-    for (size_t i = 0; i < num_stacks; ++i)
+    char *ret = malloc(num_stacks * sizeof(char));
+    for (int i = 0; i < num_stacks; ++i)
         ret[i] = stacks[i][size[i] - 1];
 
     // clear the memory
-    for (size_t i = 0; i < num_stacks; ++i)
+    for (int i = 0; i < num_stacks; ++i)
         free(stacks[i]);
     free(stacks);
     free(size);
@@ -123,23 +133,25 @@ char* one(FILE* fp)
     return ret;
 }
 
-char* two(FILE* fp)
+char *two(FILE *fp)
 {
     // prep the input
-    char* line = NULL;
+    char *line = NULL;
     ssize_t read;
     size_t len = 0;
 
     // read until the newline
     size_t des_line = 0;
     int num_stacks = 0;
-    while ((read = getline(&line, &len, fp)) != 1) {
+    while ((read = getline(&line, &len, fp)) != 1)
+    {
         // our desired line guaranteed starts with a space
-        if (line[0] == ' ') {
+        if (line[0] == ' ')
+        {
             // get the number of stacks
             size_t num_idx = 0;
-            char* num = malloc(3 * sizeof(char));
-            for (size_t idx = read - 2; line[idx] != ' '; --idx)
+            char *num = malloc(3 * sizeof(char));
+            for (int idx = read - 2; line[idx] != ' '; --idx)
                 num[num_idx++] = line[idx];
             num_stacks = atoi(num);
             break;
@@ -151,20 +163,23 @@ char* two(FILE* fp)
     fseek(fp, 0, 0);
 
     // allocate the storage
-    char** stacks = malloc(num_stacks * sizeof(char*));
-    for (size_t i = 0; i < num_stacks; ++i) {
+    char **stacks = malloc(num_stacks * sizeof(char *));
+    for (int i = 0; i < num_stacks; ++i)
+    {
         stacks[i] = malloc(100 * sizeof(char));
     }
-    size_t* size = malloc(9 * sizeof(size_t));
+    size_t *size = malloc(9 * sizeof(size_t));
 
     // now read the first 8 lines and get the data
     size_t line_num = 0;
-    while (line_num != des_line) {
+    while (line_num != des_line)
+    {
         // get the line
         read = getline(&line, &len, fp);
 
         // read across the line looking for chars
-        for (size_t i = 0; i < read; ++i) {
+        for (int i = 0; i < read; ++i)
+        {
             if (line[i] >= 'A' && line[i] <= 'Z')
                 stacks[i / 4][size[i / 4]++] = line[i];
         }
@@ -173,8 +188,9 @@ char* two(FILE* fp)
     }
 
     // we put them in backwards, reverse the lists
-    for (size_t i = 0; i < num_stacks; ++i) {
-        char* rev_stack = malloc(100 * sizeof(char));
+    for (int i = 0; i < num_stacks; ++i)
+    {
+        char *rev_stack = malloc(100 * sizeof(char));
         for (size_t j = 0; j <= size[i]; ++j)
             rev_stack[j] = stacks[i][size[i] - j - 1];
         free(stacks[i]);
@@ -187,7 +203,8 @@ char* two(FILE* fp)
     getline(&line, &len, fp);
 
     // now we need to parse the rest of the lines
-    while (getline(&line, &len, fp) != -1) {
+    while (getline(&line, &len, fp) != -1)
+    {
         // move to first number
         size_t idx;
         for (idx = 1; line[idx - 1] != ' '; ++idx)
@@ -195,7 +212,7 @@ char* two(FILE* fp)
 
         // get first number
         size_t num_idx = 0;
-        char* num = malloc(3 * sizeof(char));
+        char *num = malloc(3 * sizeof(char));
         for (; line[idx] != ' '; ++idx)
             num[num_idx++] = line[idx];
         const size_t moving = atoi(num);
@@ -225,19 +242,21 @@ char* two(FILE* fp)
         const size_t destination = atoi(num) - 1;
 
         // perform the action
-        for (size_t action = 0; action < moving; ++action) {
-            stacks[destination][size[destination]++] = stacks[origin][(size[origin] - moving) + action];
+        for (size_t action = 0; action < moving; ++action)
+        {
+            stacks[destination][size[destination]++]
+                = stacks[origin][(size[origin] - moving) + action];
         }
         size[origin] -= moving;
     }
 
     // store top stacks in return
-    char* ret = malloc(num_stacks * sizeof(char));
-    for (size_t i = 0; i < num_stacks; ++i)
+    char *ret = malloc(num_stacks * sizeof(char));
+    for (int i = 0; i < num_stacks; ++i)
         ret[i] = stacks[i][size[i] - 1];
 
     // clear the memory
-    for (size_t i = 0; i < num_stacks; ++i)
+    for (int i = 0; i < num_stacks; ++i)
         free(stacks[i]);
     free(stacks);
     free(size);
@@ -247,8 +266,9 @@ char* two(FILE* fp)
 
 int main()
 {
-    FILE* fp = fopen("day05.txt", "r");
-    if (!fp) {
+    FILE *fp = fopen("day05.txt", "r");
+    if (!fp)
+    {
         printf("Bad file read\n");
         exit(EXIT_FAILURE);
     }
